requirements.txt
paho-mqtt
pysimplegui
bluez
pyserial
dbus-python
PyGObject
bless
pytz